[CleanUp] Cleanup code of 'resize' features into 'change-bounds'
- Move files from resize to change-bounds directory
- Remove duplicate code
- Rename resizeCommandModule to changeBoundsCommandModule
