+++ client/packages/glsp-sprotty/src/features/change-bounds/resize.ts
